add TestObject
