+++ src/core/test/lib/index.js
@@ -1,0 +1,2 @@
+export { default as isTestFn } from './isTestFn';
+export { default as TestObject } from './TestObject';
