+++ src/core/test/spec.js
